requirements.txt
fastapi==0.115.0
python-dotenv==1.0.1
uvicorn==0.30.6
streamlit==1.39.0
httpx==0.27.2
beautifulsoup4==4.12.3
pydantic==2.9.2
pydantic-settings==2.5.2
pyyaml==6.0.2
redis==5.1.1
pytest==8.3.3
pytest-asyncio==0.24.0
ruff==0.6.9
mypy==1.11.2
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20241016
